Special Functions

These function are specific to cavity polariton research, where things like the cavity transmittance, material dielectric function, and cavity free spectral range (FSR) need to be calculated and modeled.

CurveFitModels.cavity_transmittance — Method

cavity_transmittance(p, ν)

Fabry-Pérot cavity transmittance with an absorbing medium as a function of frequency.

Arguments

p: Parameters [n, α, L, R, ϕ]

n: Refractive index

α: Absorption coefficient

L: Cavity length

R: Mirror reflectance (T = 1 - R assumed)

ϕ: Phase shift upon reflection

ν: Frequency (independent variable)

\[\begin{aligned} T(\nu) = \frac{(1-R)^2 e^{-\alpha L}}{1 + R^2 e^{-2\alpha L} - 2R e^{-\alpha L} \cos(4\pi n L \nu + 2\phi)} \end{aligned}\]

https://en.wikipedia.org/wiki/Fabry%E2%80%93P%C3%A9rot_interferometer

CurveFitModels.complex_lorentzian — Method

complex_lorentzian(p, ν)

Complex Lorentzian susceptibility function.

Arguments

p: Parameters [A, ν₀, Γ]

A: Oscillator strength

ν₀: Resonance frequency

Γ: Linewidth (damping)

ν: Frequency (independent variable)

\[\begin{aligned} \chi(\nu) = \frac{A}{\nu_0^2 - \nu^2 - i\Gamma\nu} \end{aligned}\]

The real and imaginary parts are dielectric_real and dielectric_imag, respectively.

https://en.wikipedia.org/wiki/Lorentzoscillatormodel

CurveFitModels.dielectric_imag — Method

dielectric_imag(p, ν)

Imaginary part of complex Lorentzian susceptibility (Lorentz oscillator model).

This is equivalent to imag(complex_lorentzian(p, ν)).

Arguments

p: Parameters [A, ν₀, Γ]

A: Oscillator strength

ν₀: Resonance frequency

Γ: Linewidth (damping)

ν: Frequency (independent variable)

\[\begin{aligned} \chi_2(\nu) = \text{Im}\left[\frac{A}{\nu_0^2 - \nu^2 - i\Gamma\nu}\right] = \frac{A \Gamma \nu}{(\nu_0^2 - \nu^2)^2 + (\Gamma\nu)^2} \end{aligned}\]

https://en.wikipedia.org/wiki/Lorentzoscillatormodel

CurveFitModels.dielectric_real — Method

dielectric_real(p, ν)

Real part of complex Lorentzian susceptibility (Lorentz oscillator model).

This is equivalent to real(complex_lorentzian(p, ν)).

Arguments

p: Parameters [A, ν₀, Γ]

A: Oscillator strength

ν₀: Resonance frequency

Γ: Linewidth (damping)

ν: Frequency (independent variable)

\[\begin{aligned} \chi_1(\nu) = \text{Re}\left[\frac{A}{\nu_0^2 - \nu^2 - i\Gamma\nu}\right] = \frac{A (\nu_0^2 - \nu^2)}{(\nu_0^2 - \nu^2)^2 + (\Gamma\nu)^2} \end{aligned}\]

https://en.wikipedia.org/wiki/Lorentzoscillatormodel